Chapter 17 Practice Identify Photo Folders
Added try and except statements to the "im = Image.open(os.path.join(foldername, filename))" line to prevent an error if the file has the .jpg or .png extension but is an empty file that Image.py is unable to identify as an image file. The except statement just continues the loop to the next file instead of stopping and presenting an OSError.